+++ includes/Adapter/internal/ConfigJson.hpp
@@ -44,18 +44,20 @@ struct TypedDecoder {
  * `json` is expected to be a JSON object with a field `"type"`. Furthermore,
  * one of the following must hold:
  *
- * - `"type"` has value `"linear"` and there are further fields `"factor"` and
- *   `"offset"` of JSON type `number`.
- *   The decoder converts the given registers to an unsigned integer in a little
- *   endian way and then applies the given linear transformation.
+ * `"type"` has value `"linear"` and there are further fields:
+ * - optionally `"signed"` of JSON type `boolean` with default `false`
+ * - optionally `"factor"` of JSON type `number` with default `1`
+ * - optionally `"offset"` of JSON type `number` with default `0`.
+ * The decoder converts the given registers to a signd or unsigned integer in a
+ * little endian way and then applies the given linear transformation.
  *
- * - `"type"` has value `"float"` and there are exactly two registers.
- *   The decoder treats the registers as an IEEE 754 Single, given in little
- *   endian.
+ * `"type"` has value `"float"` and there are exactly two registers.
+ * The decoder treats the registers as an IEEE 754 Single, given in little
+ * endian.
  *
- * - `"type"` has value `"float"` and there are exactly four registers.
- *   The decoder treats the registers as an IEEE 754 Double, given in little
- *   endian.
+ * `"type"` has value `"float"` and there are exactly four registers.
+ * The decoder treats the registers as an IEEE 754 Double, given in little
+ * endian.
  *
  * @throws `std::runtime_error
  * @throws whatever `nlohmann/json` throws